Update export pattern
