+++ src/components/GlobalStyle/index.js
@@ -1,21 +1,3 @@
-import { createGlobalStyle } from 'styled-components';
-
-const GlobalStyle = createGlobalStyle`
-  html {
-    font-size: 62.5%;
-  }
-
-  body {
-    margin: 0;
-    padding: 0;
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    overflow-y: scroll;
-    font-weight: 400;
-    font-size: 1.4rem;
-    line-height: 1.4;
-  }
-`;
+import GlobalStyle from './GlobalStyle';
 
 export default GlobalStyle;
