+++ src/components/Root/Root.js
@@ -1,0 +1,32 @@
+// Root of application
+// Home of various root providers (such as redux or router) and css global styles
+
+import '../../styles/sanitize.css';
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Provider } from 'react-redux';
+import { BrowserRouter as Router } from 'react-router-dom';
+
+import App from '../App';
+import GlobalStyle from '../GlobalStyle';
+import ErrorBoundary from '../ErrorBoundary';
+
+function Root({ store }) {
+  return (
+    <Provider store={store}>
+      <GlobalStyle />
+      <Router>
+        <ErrorBoundary>
+          <App />
+        </ErrorBoundary>
+      </Router>
+    </Provider>
+  );
+}
+
+Root.propTypes = {
+  store: PropTypes.object.isRequired
+};
+
+export default Root;
